Main Server/requirements.txt
streamlit
Flask
openai
gtts
requests
python-dotenv
pymongo
gridfs
ultralytics
numpy
opencv-python